refactor: Datenmodelle von Lukas Bugelnig (App) auf den Server portiert
- City, Continent, PlayerState Models waren ursprünglich in der App (Kotlin)
- Auf den Server (Java) portiert, da Spiellogik serverseitig sein muss
- CityDistributor für faire Städteverteilung nach Kontinenten hinzugefügt
- Alte messaging.dtos.PlayerState entfernt, durch game.models.PlayerState ersetzt
- Imports in GameCommandService, LobbyService, GameRoomState aktualisiert
- CityDistributorTest mit 6 Unit Tests hinzugefügt
