--- src/main/java/at/aau/serg/websocketdemoserver/messaging/dtos/PlayerState.java
@@ -1,14 +1,0 @@
-package at.aau.serg.websocketdemoserver.messaging.dtos;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class PlayerState {
-    private String playerId;
-    private String assignedCity;
-    private int boardPosition;
-}
